Renamed Superion -> Hyperion::Requestor
