+++ lib/hyperion.rb
@@ -11,4 +11,5 @@ require 'active_support/core_ext/module/attribute_accessors'
 
 # ensure the requirer gets everything (except hyperion_test)
 require 'hyperion/hyperion'
+require 'hyperion/requestor'
 Dir.glob(File.join(File.dirname(__FILE__), 'hyperion/types/**/*.rb')).each { |path| require_relative(path) }
