--- lib/superion.rb
@@ -1,1 +1,0 @@
-require 'hyperion/superion'
